% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentimental analysis.R
\name{GetSentiment}
\alias{GetSentiment}
\title{Sentimental Words Analysing}
\usage{
GetSentiment(url, page_range)
}
\arguments{
\item{url}{A string indicates URL of product}

\item{page_range}{Numeric number indicates a certain page's numer}
}
\value{
Number of negative words, positive word and general sentiment
}
\description{
If the number of negative words from reviews is greater than 40 percent of total sentimental words then that product is considered as 'got BAD reviews'and vice versa
}
\examples{
GetSentiment('https://www.amazon.co.uk/find-PHRL3236-Mens-Jumpers-Off-White/dp/B004I8VJ1Y?pf_rd_p=190f0313-4dcb-4239-b2ce-39f9d6bd3b5c&pd_rd_wg=NzF31&pf_rd_r=WJ0HCRPD99CTHV8XN0K0&ref_=pd_gw_unk&pd_rd_w=yoPcr&pd_rd_r=2a39592b-aee9-47b5-a65e-862257e0f4df&th=1',1:2)
}
