% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape reviews in multiple pages.R
\name{multiple_scrape}
\alias{multiple_scrape}
\title{Create a table that scrambles page numbers and randomise page to avoid being block API by Amazon
Progressing message
Take another one second break after every three scrapes}
\usage{
multiple_scrape(product_code, page_range)
}
\arguments{
\item{product_code}{A string indicates unique code of product (aka ASIN) from Amazon}

\item{page_range}{a rage of number indicating starting page and ending page to extract reviews}
}
\value{
a list of reviews from multiple pages
}
\description{
Scrape reviews from single page and loop for other page
}
