% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape reviews.R
\name{scrape_reviews}
\alias{scrape_reviews}
\title{Create URL of Amazon product's reviews page
Extract text from the given URL}
\usage{
scrape_reviews(product_code, page_num)
}
\arguments{
\item{product_code}{A string indicates unique code of product (aka ASIN) from Amazon}

\item{page_num}{numeric number indicates a certain page's numer}
}
\value{
a list of reviews from indicated page
}
\description{
Create URL of Amazon product's reviews page
Extract text from the given URL
}
