% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentimental analysis.R
\name{GetSentiment}
\alias{GetSentiment}
\title{Sentimental Words Analysing}
\usage{
GetSentiment(product_code, page_range)
}
\arguments{
\item{product_code}{A string indicates a unique code of product (aka ASIN) from Amazon}

\item{page_range}{numeric number indicates a certain page's numer}
}
\value{
Number of negative words, positive word and general sentiment
}
\description{
If the number of negative words from reviews is greater than 40 percent of total sentimental words then that product is considered as 'got BAD reviews'and vice sera
}
\examples{
GetSentiment('B004I8VJ1Y',1:2)

}
