% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentimental analysis.R
\name{GetSentiment}
\alias{GetSentiment}
\title{Sentimental words analysing}
\usage{
GetSentiment(product_code, page_range)
}
\arguments{
\item{product_code}{A string indicates unique code of product (aka ASIN) from Amazon}

\item{page_range}{numeric number indicates a certain page's numer}
}
\value{
Number of negative words, positive word and general sentiment
}
\description{
If the number of negative words is greater than 40% of total sentimental words then showing message 'got BAD reviews'  and vice sera.
}
