% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape reviews in multiple pages.R
\name{multiple_scrape}
\alias{multiple_scrape}
\title{Multiple Reviews Scraping}
\usage{
multiple_scrape(url, page_range)
}
\arguments{
\item{url}{URL of product}

\item{page_range}{a rage of number indicating starting page and ending page to extract reviews}
}
\value{
a list of reviews from multiple pages
}
\description{
Scrape reviews from a single page and loop for other page by randomising orders of pages to avoid being block API
}
\examples{
multiple_scrape('https://www.amazon.co.uk/find-PHRL3236-Mens-Jumpers-Off-White/dp/B004I8VJ1Y?pf_rd_p=190f0313-4dcb-4239-b2ce-39f9d6bd3b5c&pd_rd_wg=NzF31&pf_rd_r=WJ0HCRPD99CTHV8XN0K0&ref_=pd_gw_unk&pd_rd_w=yoPcr&pd_rd_r=2a39592b-aee9-47b5-a65e-862257e0f4df&th=1',1:4)

}
