% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbing.R
\name{plotbing}
\alias{plotbing}
\title{Sentimental Words Plotting}
\usage{
plotbing(product_code, page_range)
}
\arguments{
\item{product_code}{A string indicates a unique code of product (aka ASIN) from Amazon}

\item{page_range}{a rage of number indicating starting page and ending page to extract reviews}
}
\value{
Null
}
\description{
Plotting a column graph, shows words whose highest frequencies
}
\examples{
plotbing('B004I8VJ1Y',1:4)

}
