% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape reviews.R
\name{scrape_reviews}
\alias{scrape_reviews}
\title{Reviews Scraping}
\usage{
scrape_reviews(product_code, page_num)
}
\arguments{
\item{product_code}{A string indicates a unique code of product (aka ASIN) from Amazon}

\item{page_num}{numeric number indicates a certain page's numer}
}
\value{
a list of reviews from indicated page
}
\description{
Scrape product's reviews from a single review page
}
\examples{
 scrape_reviews('B004I8VJ1Y',10)

}
